Decouple logging config from logging module and move to starlite.config.logging
