+++ starlite/config/__init__.py
@@ -2,6 +2,7 @@ from .cache import CacheConfig
 from .compression import CompressionConfig
 from .cors import CORSConfig
 from .csrf import CSRFConfig
+from .logging import LoggingConfig
 from .openapi import OpenAPIConfig
 from .static_files import StaticFilesConfig
 from .template import TemplateConfig
@@ -14,4 +15,5 @@ __all__ = [
     "StaticFilesConfig",
     "TemplateConfig",
     "CompressionConfig",
+    "LoggingConfig",
 ]
